refactor: extract usePanelHeights hook from BottomPanel
isCollapsed was the only local-state input to getPanelHeights, with
everything else (window height, dataTableHeight, CSS vars) pulled
straight from context - move that plumbing into its own hook per
review feedback.
